# api/docs/swagger.yaml
definitions:
  content.Answer:
    properties:
      topdestinations:
        items:
          $ref: '#/definitions/content.TopDestinationsRes'
        type: array
    type: object
  content.Author:
    properties:
      full_name:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  content.CreateTipsReq:
    properties:
      category:
        type: string
      content:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
  content.CreateTipsRes:
    properties:
      author_id:
        type: string
      category:
        type: string
      content:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
  content.Destinations:
    properties:
      country:
        type: string
      currency:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  content.GetDestinationsByIdRes:
    properties:
      average_cost_per_day:
        type: string
      best_time_to_visit:
        type: string
      country:
        type: string
      currency:
        type: string
      description:
        type: string
      id:
        type: string
      language:
        type: string
      name:
        type: string
    type: object
  content.GetDestinationsRes:
    properties:
      destination:
        items:
          $ref: '#/definitions/content.Destinations'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  content.GetMessagesRes:
    properties:
      limit:
        type: integer
      messages:
        items:
          $ref: '#/definitions/content.Messages'
        type: array
      offset:
        type: integer
      total:
        type: integer
    type: object
  content.GetTipsRes:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      tips:
        items:
          $ref: '#/definitions/content.Tips'
        type: array
      total:
        type: integer
    type: object
  content.GetUserStatRes:
    properties:
      most_popular_itinerary:
        $ref: '#/definitions/content.PopularItinerary'
      most_popular_story:
        $ref: '#/definitions/content.PopularStory'
      total_comments_received:
        type: string
      total_countries_visited:
        type: string
      total_itineraries:
        type: string
      total_likes_received:
        type: string
      total_stories:
        type: string
      user_id:
        type: string
    type: object
  content.Messages:
    properties:
      content:
        type: string
      id:
        type: string
      recipient:
        $ref: '#/definitions/content.Author'
      sender:
        $ref: '#/definitions/content.Author'
    type: object
  content.PopularItinerary:
    properties:
      id:
        type: string
      likes_count:
        type: string
      title:
        type: string
    type: object
  content.PopularStory:
    properties:
      id:
        type: string
      likes_count:
        type: string
      title:
        type: string
    type: object
  content.SendMessageReq:
    properties:
      content:
        type: string
      recipient_id:
        type: string
      user_id:
        type: string
    type: object
  content.SendMessageRes:
    properties:
      content:
        type: string
      id:
        type: string
      recipient_id:
        type: string
      user_id:
        type: string
    type: object
  content.Tips:
    properties:
      author:
        $ref: '#/definitions/content.Author'
      category:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
  content.TopDestinationsRes:
    properties:
      best_time_to_visit:
        type: string
      country:
        type: string
      description:
        type: string
      popularity_score:
        type: integer
    type: object
...
info:
  contact: {}
paths:
  /api/v1/destinations:
    get:
      description: get destination
      parameters:
      - description: limit
        in: query
        name: limit
        type: string
      - description: offset
        in: query
        name: offset
        type: string
      - description: name
        in: query
        name: name
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/content.GetDestinationsRes'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: get
      tags:
      - destinations
  /api/v1/destinations/{destination_id}:
    get:
      description: Get destination by id
      parameters:
      - description: destination_id
        in: path
        name: destination_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/content.GetDestinationsByIdRes'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get
      tags:
      - destinations
...
  /api/v1/messages:
    get:
      description: get Message
      parameters:
      - description: Number of messages to fetch
        in: query
        name: limit
        type: string
      - description: Number of messages to omit
        in: query
        name: offset
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/content.GetMessagesRes'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: get Message
      tags:
      - message
    post:
      description: Send Message
      parameters:
      - description: info
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/content.SendMessageReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/content.SendMessageRes'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Send Message
      tags:
      - message
...
  /api/v1/travel-tips:
    get:
      description: get tips
      parameters:
      - description: Number of messages to fetch
        in: query
        name: limit
        type: string
      - description: Number of messages to omit
        in: query
        name: offset
        type: string
      - description: category
        in: query
        name: category
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/content.GetTipsRes'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: get
      tags:
      - tips
    post:
      description: create tips
      parameters:
      - description: destination_id
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/content.CreateTipsReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/content.CreateTipsRes'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: create
      tags:
      - tips
  /api/v1/trending-destinations:
    get:
      description: get top places
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/content.Answer'
        "500":
          description: Server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: top places
      tags:
      - top
  /api/v1/users/{user_id}/statistics:
    get:
      description: get user
      parameters:
      - description: user_id
        in: path
        name: user_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/content.GetUserStatRes'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: best
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    description: API Gateway
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"